create folder for select and input component
