+++ src/components/Form/Select/CategoryField.tsx
@@ -1,23 +1,19 @@
 import React, { useState } from 'react'
-import { getItems } from 'services/item';
-import { Item } from 'services/item/types';
-import { SelectFieldProps } from './types';
+import { Category } from 'services/category/types';
+import { getCategorys } from 'services/category';
 import { SelectField } from './SelectField';
+import { SelectFieldProps } from '../types';
 
-export const ItemField = (props: SelectFieldProps<Item>) => {
+export const CategoryField = (props: SelectFieldProps<Category>) => {
   const [params, setQueryParams] = useState({
     keywords: '',
-    orderBy: 'brand',
+    orderBy: 'name',
     orderType: 'ASC',
     page: 1,
     limit: 10,
   });
 
-  const { data } = getItems(params);
-  const onChange = (data: Item | Item[]) => {
-    props.onChange && props.onChange(data);
-  };
-
+  const { data } = getCategorys(params);
   const onInput = (keywords: string) => {
     setQueryParams(prev => ({
       ...prev,
@@ -27,15 +23,13 @@ export const ItemField = (props: SelectFieldProps<Item>) => {
 
   return (
     <>
-      <label className="form-label">
-        {props.label}
-      </label>
       <SelectField 
+        creatable={true}
+        isMulti={true}
         data={data?.data || []}
         defaultValue={props.defaultValue}
         handleInput={onInput}
-        handleChange={onChange}
-        keyLabel='brand'
+        handleChange={props.onChange}
         {...props}
       />
     </>
