+++ src/components/Form/Select/SupplierField.tsx
@@ -1,10 +1,10 @@
 import React, { useState } from 'react'
-import { Category } from 'services/category/types';
-import { getCategorys } from 'services/category';
+import { SelectFieldProps } from '../types';
+import { getSuppliers } from 'services/supplier';
+import { Supplier } from 'services/supplier/types';
 import { SelectField } from './SelectField';
-import { SelectFieldProps } from './types';
 
-export const CategoryField = (props: SelectFieldProps<Category>) => {
+export const SupplierField = (props: SelectFieldProps<Supplier>) => {
   const [params, setQueryParams] = useState({
     keywords: '',
     orderBy: 'name',
@@ -13,7 +13,10 @@ export const CategoryField = (props: SelectFieldProps<Category>) => {
     limit: 10,
   });
 
-  const { data } = getCategorys(params);
+  const { data } = getSuppliers(params);
+  const onChange = (data: Supplier | Supplier[]) => {
+    props.onChange && props.onChange(data);
+  };
   const onInput = (keywords: string) => {
     setQueryParams(prev => ({
       ...prev,
@@ -24,12 +27,10 @@ export const CategoryField = (props: SelectFieldProps<Category>) => {
   return (
     <>
       <SelectField 
-        creatable={true}
-        isMulti={true}
         data={data?.data || []}
         defaultValue={props.defaultValue}
         handleInput={onInput}
-        handleChange={props.onChange}
+        handleChange={onChange}
         {...props}
       />
     </>
